requirements.txt
streamlit==1.28.1
tensorflow==2.13.0
pillow==10.0.1
numpy==1.24.3
pandas==2.0.3
plotly==5.17.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
requests==2.31.0